backend/requirements.txt
opencv-python==4.11.0.86
deep-translator==1.11.4
pytesseract==0.3.13
openai-whisper==20240930